tests 1 and 2 pass (lilachs fbr with lfu)

+++ ex4/CacheFS.cpp
@@ -8,6 +8,7 @@
 #include "ICacheAlgorithm.hpp"
 #include "LruAlgorithm.h"
 #include "FbrAlgorithm.h"
+#include "LfuAlgorithm.h"
 #include <bitset>
 #include <cstring>
 #include <unordered_set>
@@ -130,20 +131,20 @@ int CacheFS_pread(int file_id, void *buf, size_t count, off_t offset)
     size_t buf_offset = 0;
     size_t addToOffset = 0;
 
-    long toRead = std::min(fsize-junkBits, count);
+    long toRead = std::min(fsize - junkBits, count);
 
     long cur_count = count;
     int blockCandid = (int) (offset / blockSize);
     void *_cacheBuff;
-    while (buf_offset<toRead && cur_count>0)
+    while (buf_offset < (size_t) toRead && cur_count > 0)
     {
         std::pair<std::string, int> key = std::make_pair(_openFiles[file_id], blockCandid);
         if ((_cacheBuff = _algorithm->Get(key)) != nullptr)
         {
             cache_hits++;
 
-            addToOffset = std::min(strlen((char*)_cacheBuff) -junkBits, (size_t) cur_count);
-            addToOffset = std::min(addToOffset,toRead-buf_offset);
+            addToOffset = std::min(strlen((char *) _cacheBuff) - junkBits, (size_t) cur_count);
+            addToOffset = std::min(addToOffset, toRead - buf_offset);
 //
 //            size_t blockSize = GetBlockSize();
 //    size_t junkBits = offset % blockSize;
@@ -182,7 +183,7 @@ int CacheFS_pread(int file_id, void *buf, size_t count, off_t offset)
             }
             cache_misses++;
             _algorithm->Set(key, _cacheBuff);
-            addToOffset = std::min(_readSize - junkBits, (size_t)cur_count);
+            addToOffset = std::min(_readSize - junkBits, (size_t) cur_count);
             memcpy((char *) buf + buf_offset, (char *) _cacheBuff + junkBits, addToOffset);
         }
         buf_offset += addToOffset;
